+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.5.2
-* add: `react-ts`
+* add: `react-ts` beta
+* add: choose add `--global`
 ## 1.5.1
 * fix:  eslint rule `no-unused-vars` not work in `vue2` `vue3`
 * add: support `react`